Basic folder restructure.
