+++ packages/react-components/src/Maps/AnimationControl/AnimationControl.tsx
