+++ packages/react-components/src/Maps/AnimationLayer/helpers.ts
